enrich api response  msg (api/UserApiController.java), update progress

+++ SpringTwitter/src/main/java/com/yen/tweet/api/UserApiController.java
@@ -1,6 +1,6 @@
 package com.yen.tweet.api;
 
-// p.137  p.138
+// p.137 p.138 p.141
 
 /*
  * THE CONTROLLER IMPLEMENTS CRUD OPERATION VIA RESTFUL API
@@ -12,6 +12,10 @@ import com.yen.tweet.user.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.*;
+
 import java.util.List;
 
 @RestController
@@ -20,7 +24,7 @@ public class UserApiController {
 
     private UserRepository userRepository;
 
-    // constructor ?
+    // constructor
     @Autowired
     public UserApiController(UserRepository userRepository) {
         this.userRepository = userRepository;
@@ -28,21 +32,41 @@ public class UserApiController {
 
     @RequestMapping(value = "/users", method = RequestMethod.GET)
     public List<User> findAll() {
+
         return userRepository.findALL();
     }
 
     @RequestMapping(value = "/users", method = RequestMethod.POST)
-    public User createUser(@RequestBody User user) {
-        return userRepository.save(user);
+    public ResponseEntity<User> createUser(@RequestBody User user) {
+        HttpStatus status = HttpStatus.OK;
+
+        if (!userRepository.exists(user.getEmail())){
+            status = HttpStatus.CREATED;
+        }
+
+        User saved = userRepository.save(user);
+        return new ResponseEntity<>(saved, status);
     }
 
     @RequestMapping(value = "/user/{email}", method = RequestMethod.PUT)
-    public User updateUser(@PathVariable String email, @RequestBody User user){
-        return userRepository.save(email, user);
+    public ResponseEntity<User> updateUser(@PathVariable String email, @RequestBody User user){
+
+        if (!userRepository.exists(user.getEmail())){
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
+
+        User saved = userRepository.save(email, user);
+        return new ResponseEntity<>(saved, HttpStatus.CREATED);
     }
 
     @RequestMapping(value = "/user/{email}", method = RequestMethod.DELETE)
-    public void deleteUser(@PathVariable String email){
+    public ResponseEntity<User> deleteUser(@PathVariable String email){
+
+        if (!userRepository.exists(email)){
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
+
         userRepository.delete(email);
+        return new ResponseEntity<>(HttpStatus.OK);
     }
 }
